Task3(new) malyshev

+++ src/main/java/org/example/bot/processor/share/ShareUnsubscribeTestProcessor.java
@@ -1,6 +1,8 @@
-package org.example.bot.processor;
+package org.example.bot.processor.share;
 
 import org.example.bot.entity.TestEntity;
+import org.example.bot.entity.UserEntity;
+import org.example.bot.processor.AbstractCallbackProcessor;
 import org.example.bot.service.TestService;
 import org.example.bot.service.UserService;
 import org.example.bot.telegram.BotResponse;
@@ -14,13 +16,20 @@ import java.util.Optional;
  */
 @Component
 public class ShareUnsubscribeTestProcessor extends AbstractCallbackProcessor {
+    /**
+     * Сервис для работы с пользователями
+     */
     private final UserService userService;
+
+    /**
+     * Сервис для работы с тестами
+     */
     private final TestService testService;
 
     /**
      * Конструктор для инициализации обработчика callback.
      */
-    protected ShareUnsubscribeTestProcessor(UserService userService, TestService testService) {
+    public ShareUnsubscribeTestProcessor(UserService userService, TestService testService) {
         super("SHARE_UNSUBSCRIBE_TEST");
         this.userService = userService;
         this.testService = testService;
@@ -28,15 +37,21 @@ public class ShareUnsubscribeTestProcessor extends AbstractCallbackProcessor {
 
     @Override
     public BotResponse process(Long userId, String message) {
-        Long testId = Long.parseLong(extractData(message));
+        Long testId = Long.parseLong(message.split(" ")[1]);
         List<TestEntity> receivedTests = userService.getOpenReceivedTests(userId);
         Optional<TestEntity> testOpt = testService.getTest(testId);
         if(testOpt.isEmpty() || !receivedTests.contains(testOpt.get()))
             return new BotResponse("У вас нет доступа к тесту");
         TestEntity test = testOpt.get();
 
+        Optional<UserEntity> userOpt = userService.getUserById(userId);
+        if (userOpt.isEmpty())
+            return new BotResponse("Пользователь не найден");
+
+        UserEntity user = userOpt.get();
         userService.removeReceivedTest(userId, test);
-        test.getRecipients().remove(userService.getUserById(userId));
+        test.getRecipients().remove(user);
+        testService.update(test);
         String creatorUsername = userService.getUserById(test.getCreatorId()).get().getUsername();
         return new BotResponse(
                 ("Вы отписались от теста “%s (%s)” " +
